+++ client/src/components/dataset/sidebars/Filter.jsx
@@ -57,10 +57,6 @@ export default class Filter extends Component {
     return (
       <div
         className="DataTableSidebar"
-        style={{
-          width: '300px',
-          height: 'calc(100vh - 4rem)',
-        }}
       >
         <SidebarHeader onClose={onClose}>
           Filter column {column.get('title')}